+++ lib/libesp32/Berry/default/berry_conf.h
@@ -158,6 +158,7 @@
 #define BE_USE_DEBUG_MODULE             1
 #define BE_USE_GC_MODULE                1
 #define BE_USE_SOLIDIFY_MODULE          1
+#define BE_USE_INTROSPECT_MODULE        1
 
 /* Macro: BE_EXPLICIT_XXX
  * If these macros are defined, the corresponding function will